# packages/contracts/openapi/private-api.v1.json
{
  "components": {
    "schemas": {
...
      "DomainCreate": {
        "additionalProperties": false,
        "properties": {
          "domain": {
            "maxLength": 253,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          }
        },
        "required": [
          "domain"
        ],
        "title": "DomainCreate",
        "type": "object"
      },
      "DomainResponse": {
        "additionalProperties": false,
        "properties": {
          "canonical_name": {
            "title": "Canonical Name",
            "type": "string"
          },
          "contract_version": {
            "const": "v1",
            "default": "v1",
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "ownership_state": {
            "enum": [
              "pending",
              "verified",
              "expired",
              "failed",
              "revoked",
              "reverification_required"
            ],
            "title": "Ownership State",
            "type": "string"
          },
          "registrable_domain": {
            "title": "Registrable Domain",
            "type": "string"
          },
          "unicode_display": {
            "title": "Unicode Display",
            "type": "string"
          },
          "warnings": {
            "items": {
              "enum": [
                "idn_present",
                "mixed_scripts"
              ],
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "organization_id",
          "canonical_name",
          "unicode_display",
          "registrable_domain",
          "warnings",
          "ownership_state",
          "created_at"
        ],
        "title": "DomainResponse",
        "type": "object"
      },
      "EmergencyControlCreate": {
        "additionalProperties": false,
        "properties": {
          "domain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "operation_class": {
            "anyOf": [
              {
                "enum": [
                  "dns_verification",
                  "https_verification",
                  "passive_assessment",
                  "active_assessment"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Class"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "scope": {
            "enum": [
              "global",
              "organization",
              "domain",
              "operation_class"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "reason"
        ],
        "title": "EmergencyControlCreate",
        "type": "object"
      },
      "EmergencyControlDeactivate": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "EmergencyControlDeactivate",
        "type": "object"
      },
      "EmergencyControlResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "contract_version": {
            "const": "v1",
            "default": "v1",
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "operation_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Class"
          },
          "organization_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "scope": {
            "enum": [
              "global",
              "organization",
              "domain",
              "operation_class"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "reason",
          "active",
          "created_at"
        ],
        "title": "EmergencyControlResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SIEMBIOT Private API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/emergency-controls": {
      "post": {
        "operationId": "activate_global_api_v1_emergency_controls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyControlCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Global",
        "tags": [
          "emergency-controls"
        ]
      }
    },
    "/api/v1/emergency-controls/{control_id}/deactivate": {
      "post": {
        "operationId": "deactivate_global_api_v1_emergency_controls__control_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyControlDeactivate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate Global",
        "tags": [
          "emergency-controls"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_id}/domains": {
      "get": {
        "operationId": "list_domains_api_v1_organizations__organization_id__domains_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainResponse"
                  },
                  "title": "Response List Domains Api V1 Organizations  Organization Id  Domains Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Domains",
        "tags": [
          "domains"
        ]
      },
      "post": {
        "operationId": "create_domain_api_v1_organizations__organization_id__domains_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Domain",
        "tags": [
          "domains"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/domains/{domain_id}": {
      "get": {
        "operationId": "get_domain_api_v1_organizations__organization_id__domains__domain_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Domain Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Domain",
        "tags": [
          "domains"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_id}/emergency-controls": {
      "get": {
        "operationId": "list_controls_api_v1_organizations__organization_id__emergency_controls_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmergencyControlResponse"
                  },
                  "title": "Response List Controls Api V1 Organizations  Organization Id  Emergency Controls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Controls",
        "tags": [
          "emergency-controls"
        ]
      },
      "post": {
        "operationId": "activate_api_v1_organizations__organization_id__emergency_controls_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyControlCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate",
        "tags": [
          "emergency-controls"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/emergency-controls/{control_id}/deactivate": {
      "post": {
        "operationId": "deactivate_api_v1_organizations__organization_id__emergency_controls__control_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyControlDeactivate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyControlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate",
        "tags": [
          "emergency-controls"
        ]
      }
    },
...
  }
}